% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{add_attachment}
\alias{add_attachment}
\title{Add an attachment to your email. This can be any filetype.}
\usage{
add_attachment(filepath)
}
\arguments{
\item{filepath}{Path to the file you'd like to attach. *Note* that this will
be the filepath *in your bandit job!*.}
}
\description{
Add an attachment to your email. This can be any filetype.
}
\examples{
\dontrun{
add_attachment("my-plot.png")
add_attachment("train.csv")
add_attachment("scored-data.csv")
}
}
