% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R
\docType{class}
\name{bandit-class}
\alias{bandit}
\alias{bandit-class}
\title{A bandit reference class (RC) object.}
\description{
A bandit reference class (RC) object.
}
\section{Fields}{

\describe{
\item{\code{samples}}{a tibble containing samples.}

\item{\code{models}}{a tibble containing trained models.}

\item{\code{jobs}}{a tibble containing completed jobs.}

\item{\code{\code{alpha}}}{the linear UCB tuning parameter.}

\item{\code{\code{formula}}}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model that is fitted.}

\item{\code{\code{family}}}{supported response type}

\item{\code{\code{contrasts}}}{an optional list. See the \code{contrasts.arg} of \code{\link{model.matrix.default}}.}

\item{\code{\code{xlevels}}}{(where relevant) a record of the levels of the factors used in fitting.}

\item{\code{\code{currentJob}}}{a numeric with the id of the current job.}

\item{\code{\code{currentModel}}}{the prototype of the latest fit.}

\item{\code{\code{currentParams}}}{a list of tuning parameters.}

\item{\code{\code{variables}}}{a character vectors of variables extracted from the bandit formula.}
}}
\section{Methods}{

\describe{
\item{\code{addSamples(df)}}{add samples to the bandit. \code{df} is coercible to a data.frame, and can be appended
to the \code{data.frame} used at creation. In particular, it contains an \code{id} column
that is a primary key.}

\item{\code{initialize(formula, data, alpha = 1, contrasts = NULL, ...)}}{initialization function}

\item{\code{train(data, seed = NULL, parRidge = NULL, parLasso = NULL)}}{train the bandit using available samples and latest values of the tuning
parameters.}

\item{\code{tune(data, param = "lambdaRidge", value = "auto",
  lambdaAuto = "lambda.1se", parCvGlmnet = NULL, seed = NULL)}}{set the value of a tuning parameter of the bandit. \code{param} is either
\code{'lambdaRidge'} (the default) or \code{'lambdaLasso'}. Setting
\code{lambdaLasso} > 0 enables a first stage of variable selection.
\code{value} is either a scalar in [0,1] or 'auto'. If
set to \code{'auto'}, the parameter is picked using \code{\link[glmnet]{cv.glmnet}}.
\code{lambdaAuto} is either \code{'lambda.1se'} or \code{'lambda.min'}
depending on which outcome of \code{\link[glmnet]{cv.glmnet}} should be selected. It
is ignored if \code{value} is not \code{'auto'}. \code{parCvGlmnet} is a list of
parameters passed to \code{\link[glmnet]{cv.glmnet}}. This parameter is ignored if
\code{value} is not \code{'auto'}.}
}}

