% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditThompson.R
\docType{class}
\name{bandit_thompson-class}
\alias{bandit_thompson-class}
\alias{bandit_thompson}
\title{A Thompson sampling bandit reference class (RC) object.}
\description{
A Thompson sampling bandit reference class (RC) object.
}
\section{Fields}{

\describe{
\item{\code{\code{gamma}}}{the blabla tuning parameter. See details.}
}}

\section{Methods}{

\describe{
\item{\code{initialize(formula, data, family, contrasts = NULL, newLevels = FALSE,
  db = NULL, path = NULL, ...)}}{initialization function}

\item{\code{train(FUN, args, seed)}}{train the bandit using available samples and latest values of the tuning
parameters.}

\item{\code{tune(FUN, args, seed)}}{set the value of a tuning parameter of the bandit. \code{param} is either
\code{'lambdaRidge'} (the default) or \code{'lambdaLasso'}. Setting
\code{lambdaLasso} > 0 enables a first stage of variable selection.
\code{value} is either a scalar in [0,1] or 'auto'. If
set to \code{'auto'}, the parameter is picked using \code{\link[glmnet]{cv.glmnet}}.
\code{lambdaAuto} is either \code{'lambda.1se'} or \code{'lambda.min'}
depending on which outcome of \code{\link[glmnet]{cv.glmnet}} should be selected. It
is ignored if \code{value} is not \code{'auto'}. \code{parCvGlmnet} is a list of
parameters passed to \code{\link[glmnet]{cv.glmnet}}. This parameter is ignored if
\code{value} is not \code{'auto'}.}
}}

