% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditUcb.R
\docType{class}
\name{bandit_ucb-class}
\alias{bandit_ucb-class}
\alias{bandit_ucb}
\title{An UCB bandit reference class (RC) object.}
\usage{
bandit_ucb(formula, data, family = c("gaussian", "binomial"),
           alpha = 1, contrasts = NULL, newLevels = FALSE,
           db = NULL, path = NULL)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model that is fitted. The response
must be named \code{y}}

\item{data}{a data frame (or object coercible by \code{\link{as.data.frame}} to a data frame)
containing the variables in the model. \code{data} must contain a column named
\code{id} that uniquely identifies each observation, and a column named \code{y} that
contains the model response.}

\item{family}{a character string describing the error distribtion and link function
to be used in the model. Can be either \code{"binomial"} or \code{"gaussian"} (the default).}

\item{alpha}{the LinUCB tuning parameter. A positive scalar. Higher values of
\code{alpha} favor exploration over exploitation.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of \code{model.matrix.default}.}

\item{newLevels}{a logical value indicating whether to allow for new factor levels when
adding samples. Default is FALSE.}

\item{db}{an optional named list of arguments passed to \code{\link[RODBC]{odbcDriverConnect}}.}

\item{path}{an optional character string naming a folder open for writing.}
}
\description{
An UCB bandit reference class (RC) object.
}
\details{
The introductory vignette provides a detailed
}
\section{Fields}{

\describe{
\item{\code{alpha}}{the linear UCB tuning parameter. See details.}

\item{\code{family}}{supported response type. See \code{\link{banditGlmnet}} for details.}
}}


\section{Methods}{

\code{train(parRidge = NULL, parLasso = NULL, seed = NULL)} trains the model using
\code{\link{banditGlmnet}} and all completed experiments.
\code{parRidge} and \code{parLasso} are optional lists of
parameters passed to \code{\link{banditGlmnet}} to control the fitting process.
\code{seed} is an optional seeding value for the random number generator.

\code{tune(param = 'lambdaRidge', value = 'auto', lambdaAuto = 'lambda.1se',
parCvGlmnet = NULL, seed = NULL)} set the value of a tuning parameter of the bandit. \code{param} is either
\code{'lambdaRidge'} (the default) or \code{'lambdaLasso'}. Setting
\code{lambdaLasso} > 0 enables a first stage of variable selection.
\code{value} is either a scalar in [0,1] or 'auto'. If
set to \code{'auto'}, the parameter is picked using \code{\link[glmnet]{cv.glmnet}}.
\code{lambdaAuto} is either \code{'lambda.1se'} or \code{'lambda.min'}
depending on which outcome of \code{\link[glmnet]{cv.glmnet}} should be selected. It
is ignored if \code{value} is not \code{'auto'}. \code{parCvGlmnet} is a list of
parameters passed to \code{\link[glmnet]{cv.glmnet}}. This parameter is ignored if
\code{value} is not \code{'auto'}.

\code{addSamples(df)} add samples to the bandit. \code{df} is coercible to
a data.frame, and can be appended to the \code{data.frame} used at creation.
In particular, it contains an \code{id} column that is a primary key.

\code{addOutcomes(y)} add outcomes to the bandit. \code{y} is a named vector
whose names are samples ids.

\code{undo()} cancel the last job.
}

