% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditGlmnet.R
\name{banditGlmnet}
\alias{banditGlmnet}
\title{Estimate a banditGlmnet model}
\usage{
banditGlmnet(formula, family = c("gaussian", "binomial"), lambdaRidge,
  lambdaLasso, data, weights, subset, na.action, offset, seed = NULL,
  parRidge = NULL, parLasso = NULL, model = TRUE, method = "glmnet",
  x = FALSE, y = TRUE, contrasts = NULL)
}
\arguments{
\item{family}{supported response type. Currently, either
\code{"gaussian"} (the default) or \code{"binomial"}.}

\item{lambdaRidge}{a scalar in [0,1]. Tuning parameter of the ridge regression.}

\item{lambdaLasso}{a scalar in [0,1]. Tuning parameter of the LASSO.}

\item{seed}{an (optional) seeding number.}

\item{parRidge}{an (optional) list of parameters passed on to glmnet for the ridge
stage. Format argument = value.}

\item{parLasso}{an (optional) list of parameters passed on to glmnet for the LASSO
stage. Format argument = value}

\item{method}{the method to be used in fitting the model.
The default \code{"glmnet"} uses \code{\link[glmnet]{glmnet}};
the alternative \code{"model.frame"} returns the model frame and does
no fitting.}
}
\value{
a banditGlmnet object.
}
\description{
The syntax follows closely that of \code{\link[stats]{glm}}.
All parameters not defined on this page are defined as in
\code{\link[stats]{glm}}.
}
