% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditr.R
\docType{package}
\name{banditr}
\alias{banditr}
\alias{banditr-package}
\title{banditr: R Client for the Bandit API}
\description{
The banditr package provides functions for saving metadata, sending email, 
spawning jobs, and streaming metrics.
}
\section{Email functions}{

\itemize{
  \item \code{\link{email}}
  \item \code{\link{add_attachment}}
}
}

\section{Metadata functions}{

\itemize{
  \item \code{\link{stream}}
}
}

\section{Jobs functions}{

\itemize{
  \item \code{\link{run}}
}
}

