% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.bandit}
\alias{coef.bandit}
\title{Extract coefficients from a bandit object}
\usage{
coef.bandit(object, what = "last")
}
\arguments{
\item{object}{an object inheriting from class \code{"bandit"}}

\item{what}{either \code{"last"} (the default) for the coefficients of the model estimated
in the last training job, or \code{"all"} for the coefficients of all
training jobs.}
}
\value{
a numeric vector if \code{what = "last"}; a matrix if \code{what = "all"}.
}
\description{
Extract coefficients from a bandit object
}
