% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{email}
\alias{email}
\title{Create an email to send after your job has finished. Bandit will
automatically send this email for you.}
\usage{
email(recipients, subject, body)
}
\arguments{
\item{recipients}{Character vector of email addresses.}

\item{subject}{Subject line of the email.}

\item{body}{Body of the email. Can be raw text or HTML. This needs to be an
atomic character vector (length 1).}
}
\description{
Create an email to send after your job has finished. Bandit will
automatically send this email for you.
}
\examples{
\dontrun{
email("hi@yhat.com", "This is my Subject", "And you're seeing this body!")
email(c("paul@yhathq.com", "hi@yhat.com"), "Two People", "Will get this!")
email(c("smug@smugdouglas.com"), "HTML", "<p>Wow! HTML is <b>fun!</b></p>")
}
}
