% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditMethods.R
\name{formula.bandit}
\alias{formula.bandit}
\title{Extracting the model formula used by some model in the bandit.}
\usage{
formula.bandit(x, modelId = NULL, reduced = FALSE)
}
\arguments{
\item{x}{a bandit object}

\item{modelId}{(optional) the id of a training job.}

\item{reduced}{(optional) logical; if TRUE and some variable selection was used, the formula
of the reduced model is returned.}
}
\value{
by default, the bandit's formula.
variable selection occurs (lambda.lasso > 0), the formula for the ridge stage only
contains the variables that were selected.
}
\description{
Extracting the model formula used by some model in the bandit.
}

