% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_connection}
\alias{get_connection}
\title{Get a pre-defined database connection from Bandit.}
\usage{
get_connection(name)
}
\arguments{
\item{name}{Name of the database you'd like to get the connection for.}
}
\value{
Database connection string.
}
\description{
Get a pre-defined database connection from Bandit.
}
\examples{
get_connection('mysql-prod')
get_connection('dw')
}
