% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getSamples}
\alias{getSamples}
\alias{getJobs}
\alias{getModel}
\title{Extract quantities from a bandit object}
\usage{
getSamples(object, what = "current")

getJobs(object, what = "all")

getModel(object, what = "last")
}
\arguments{
\item{object}{an object inheriting from class \code{"bandit"}}

\item{what}{the specific object to select. See details.}
}
\value{
For \code{getSamples} and \code{getJobs}, a \code{data.frame}. For \code{getModel},
a model object.
}
\description{
Extract quantities from a bandit object
}
\details{
By default, \code{getSamples} extracts all samples
with a non-missing outcome. Alternatives are \code{"last"},
for the samples used in the last training job; \code{"remaining"} for all samples
with a missing outcome.

By default, \code{getJobs} extracts a summary of all jobs done by the bandit. The
alternative \code{"last"} extracts a summary of the last job done by the bandit.

By default, \code{getModel} extracts the model estimated at the last training job.
The id of a training job may be supplied.
}
