% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Track and associate metadata key/values with your Bandit job. These will be
stored on the Bandit server and displayed in the UI.}
\usage{
metadata(key, value)
}
\arguments{
\item{key}{Name for the metric you'd like to track (i.e. accuracy or training_size)}

\item{value}{Number/String/Character you'd like to track. (i.e. 0.84 or 250)}
}
\description{
Track and associate metadata key/values with your Bandit job. These will be
stored on the Bandit server and displayed in the UI.
}
\examples{
\dontrun{
metadata('ROC', 0.74)
metadata('accuracy', 0.97)
metadata('foo', 'bar')
}
}
