% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditMethods.R
\name{model.frame.bandit}
\alias{model.frame.bandit}
\title{Extracting the model frame used by some model in the bandit.}
\usage{
model.frame.bandit(formula, modelId = NULL, reduced = FALSE, data = NULL)
}
\arguments{
\item{modelId}{(optional) the id of a training job.}

\item{reduced}{(optional) logical; if TRUE and some variable selection was used, the model.frame
of the reduced model is returned.}

\item{data}{an optional data.frame, list or environment
(or object coercible by as.data.frame to a data.frame),
containing the variables in formula. Neither a matrix nor an array will be accepted.}

\item{object}{a bandit object}
}
\value{
by default, the model.frame used in the last model.
}
\description{
Extracting the model frame used by some model in the bandit.
}

