% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditMethods.R
\name{model.matrix.bandit}
\alias{model.matrix.bandit}
\title{Extracts a design (or model) matrix, e.g., by expanding factors to a set
of dummary variables (depending on the contrasts) and expanding interactions
similarly.}
\usage{
model.matrix.bandit(object, modelId = NULL, reduced = FALSE, data = NULL)
}
\arguments{
\item{object}{a bandit object}

\item{modelId}{the id of a training job.}

\item{data}{an optional data.frame, list or environment
(or object coercible by as.data.frame to a data.frame),
containing the variables in formula. Neither a matrix nor an array will be accepted.}

\item{stage}{a character in 'ridge', 'lasso' for the stage of the desired model.}
}
\value{
by default, the design matrix used in the last model, for the ridge stage.
}
\description{
Extracts a design (or model) matrix, e.g., by expanding factors to a set
of dummary variables (depending on the contrasts) and expanding interactions
similarly.
}

