% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bandit}
\alias{plot.bandit}
\title{Plot diagnostics for a bandit object}
\usage{
plot.bandit(x, what = "reward", ...)
}
\arguments{
\item{x}{an object inheriting from class \code{"bandit"}}

\item{what}{the type of plot. Either \code{"reward"} (the default) or \code{"coef"}.}

\item{...}{additional arguments passed to the plotting function.}
}
\description{
Two plots (selectable by \code{what}) are currently available: a plot of the reward
over time, and a plot of model coefficients over time.
}
\details{
The \code{plot.bandit} method calls other plotting functions as specified by the
parameter \code{what}. All plotting functions are prefixed with \code{"plot"}, and
have a \code{data} parameter that returns the raw data used for plotting when
set to \code{TRUE}.
}
\seealso{
\code{\link{plotReward}}, \code{\link{plotCoef}}
}
