% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCoef}
\alias{plotCoef}
\title{Plot the reward of a bandit object}
\usage{
plotCoef(object, data = FALSE, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"bandit"}}

\item{data}{logical: should the raw plot data be returned?}

\item{...}{additional parameters passed to \code{\link[graphics]{plot.default}}}
}
\description{
Plot the reward of a bandit object
}
