% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotReward}
\alias{plotReward}
\title{Plot the reward of a bandit object}
\usage{
plotReward(object, data = FALSE, cumulative = TRUE, expectedReward = TRUE,
  maxReward, ...)
}
\arguments{
\item{object}{an object inheriting from class \code{"bandit"}}

\item{data}{logical: should the raw plot data be returned?}

\item{cumulative}{logical: should the cumulative reward be plotted?}

\item{expectedReward}{logical: should the expected reward be plotted?}

\item{maxReward}{logical: should the maximum reward be plotted? (only supported for
binomial models).}

\item{...}{additional parameters passed to \code{\link[graphics]{plot.default}}}
}
\description{
Plot the reward of a bandit object
}
