% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditMethods.R
\name{predict.bandit}
\alias{predict.bandit}
\title{Predicted values based on a model in the bandit.}
\usage{
predict.bandit(object, samplesId = NULL, modelId = NULL,
  type = c("response", "uncertainty", "score"), robust = TRUE)
}
\arguments{
\item{object}{a bandit object}

\item{samplesId}{(optional) the ids of some samples}

\item{modelId}{(optional) the id of a training job.}

\item{type}{a character containing one or more of the following:
"link","response","class","uncertainty","score"}

\item{robust}{for type = uncertainty, should robust estimation be used?}

\item{stage}{a character in 'ridge', 'lasso' for the stage of the desired model.}
}
\value{
a vector of predicted values, if only one type is selected. Otherwise, a
data.frame of predicted values. If a data.frame is returned, row.names refer to the
ids of samples. The default returns the predicted values for all the samples for
which the outcome is missing, using the last model.
}
\description{
Predicted values based on a model in the bandit.
}

