% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{run}
\alias{run}
\title{Trigger a job to run on Bandit.}
\usage{
run(project, jobname)
}
\arguments{
\item{project}{Name of the job's project.}

\item{jobname}{Name of the job you'd like to run.}
}
\value{
Response from the Bandit API.
}
\description{
Trigger a job to run on Bandit.
}
\examples{
\dontrun{
run('My-First-Project', 'theFirstJob')
run('bandit-churn', 'score-leads')
run('bandit-churn', 'update-database')
}
}
