% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{stream}
\alias{stream}
\title{Stream metrics back to Bandit.}
\usage{
stream(tag, value)
}
\arguments{
\item{tag}{Identifier for the metric.}

\item{value}{A number.}
}
\value{
Response from the Bandit API.
}
\description{
Stream metrics back to Bandit.
}
\examples{
\dontrun{
stream('x', 1)
stream('x', 2)
stream('x', 3)
stream('alpha', 100)
stream('zalpha', 20.5)
}
}
