% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R
\docType{class}
\name{bandit-class}
\alias{bandit-class}
\alias{bandit}
\title{A bandit reference class (RC) object.}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model that is fitted. The response
must be named \code{y}}

\item{data}{a data frame (or object coercible by \code{\link{as.data.frame}} to a data frame)
containing the variables in the model. \code{data} must contain a column named
\code{id} that uniquely identifies each observation, and a column named \code{y} that
contains the model response.}

\item{family}{a character string describing the error distribtion and link function
to be used in the model. Can be either \code{"binomial"} or \code{"gaussian"} (the default).}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of \code{model.matrix.default}.}

\item{newLevels}{a logical value indicating whether to allow for new factor levels when
adding samples. Default is FALSE.}

\item{db}{an optional named list of arguments passed to \code{\link[RODBC]{odbcDriverConnect}}.}

\item{path}{an optional character string naming a folder open for writing.}
}
\description{
A bandit reference class (RC) object.
}
\details{
The RC \code{"bandit"} class is the main object defined in
 \code{banditr}. It allows to manage bandit experiments easily. Two
 classes inherit from \code{"bandit"}: the \code{"\link{bandit_ucb}"}
 class for LinUCB algorithms, and the \code{"\link{bandit_thompson}"} for
 Thompson sampling algorithms.

 The introductory vignette provides a detailed explanation of bandit algorithms, and
their implementation with \code{banditr}. See the Examples section.
}
\section{Fields}{

\describe{
\item{\code{formula}}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model that is fitted.}

\item{\code{contrasts}}{an optional list. See the \code{contrasts.arg} of \code{\link{model.matrix.default}}.}

\item{\code{newLevels}}{a logical value indicating whether to allow for new factor levels when adding samples.}

\item{\code{xlevels}}{if \code{newLevels} is false, a record of the levels of the factors used in fitting.}

\item{\code{currentJob}}{a numeric with the id of the current job.}

\item{\code{currentModel}}{the prototype of the latest fit.}

\item{\code{currentParams}}{a list of tuning parameters.}

\item{\code{banditData}}{the bandit's data.}

\item{\code{statistics}}{a character vector of statistics to be computed when adding outcomes}
}}


\examples{
vignette("introduction", "banditr")
}
\seealso{
\code{\link{bandit_ucb}}, \code{\link{bandit_thompson}}
}
