% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditThompson.R
\docType{class}
\name{bandit_thompson-class}
\alias{bandit_thompson-class}
\alias{bandit_thompson}
\alias{bandit_stan_lm-class}
\alias{bandit_stan_lm}
\alias{bandit_stan_glm-class}
\alias{bandit_stan_glm}
\alias{bandit_stan_glmer-class}
\alias{bandit_stan_glmer}
\title{A Thompson sampling bandit reference class (RC) object.}
\usage{
bandit_stan_lm(formula, data, gamma = 1,
               contrasts = NULL, newLevels = FALSE,
               db = NULL, path = NULL)
bandit_stan_glm(formula, data, family = c("gaussian", "binomial"),
                gamma = 1, contrasts = NULL, newLevels = FALSE,
                db = NULL, path = NULL)
bandit_stan_glmer(formula, data, family = c("gaussian", "binomial"),
                  gamma = 1, contrasts = NULL, newLevels = FALSE,
                  db = NULL, path = NULL)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model that is fitted. The response
must be named \code{y}}

\item{data}{a data frame (or object coercible by \code{\link{as.data.frame}} to a data frame)
containing the variables in the model. \code{data} must contain a column named
\code{id} that uniquely identifies each observation, and a column named \code{y} that
contains the model response.}

\item{family}{a character string describing the error distribtion and link function
to be used in the model. Can be either \code{"binomial"} or \code{"gaussian"} (the default).}

\item{gamma}{the Thompson sampling tuning parameter. See details.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of \code{model.matrix.default}.}

\item{newLevels}{a logical value indicating whether to allow for new factor levels when
adding samples. Default is FALSE.}

\item{db}{an optional named list of arguments passed to \code{\link[RODBC]{odbcDriverConnect}}.}

\item{path}{an optional character string naming a folder open for writing.}
}
\description{
A Thompson sampling bandit reference class (RC) object.
}
\details{
abc
}
\section{Fields}{

\describe{
\item{\code{gamma}}{the Thompson sampling tuning parameter.}
}}


\section{Methods}{


\code{train(..., seed = NULL)} train the model using the relevant function from
\code{\link{[rstanarm]rstanarm}} and all completed experiments. \code{...} are
additional parameters passed on to the relevant function in \code{rstanarm}.
\code{seed} is an optional seeding value for the random number generator.

\code{tune()} currently not supported.

\code{addSamples(df)} add samples to the bandit. \code{df} is coercible to
a data.frame, and can be appended to the \code{data.frame} used at creation.
In particular, it contains an \code{id} column that is a primary key.

\code{addOutcomes(y)} add outcomes to the bandit. \code{y} is a named vector
whose names are samples ids.

\code{undo()} cancel the last job.
}

