% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditGlmnet.R
\name{banditGlmnet}
\alias{banditGlmnet}
\title{Estimate a banditGlmnet model}
\usage{
banditGlmnet(formula, family = c("gaussian", "binomial"), lambdaRidge,
  lambdaLasso, data, weights, subset, na.action, offset, seed = NULL,
  parRidge = NULL, parLasso = NULL, model = TRUE, method = "glmnet",
  x = FALSE, y = TRUE, contrasts = NULL)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced
to that class): a symbolic description of the model to be fitted.}

\item{family}{supported response type}

\item{lambdaRidge}{a scalar in [0,1]. Tuning parameter of the ridge regression.}

\item{lambdaLasso}{a scalar in [0,1]. Tuning parameter of the LASSO.}

\item{data}{a data frame, list or environment (or object coercible by as.data.frame
to a data frame) containing the variables in the model.}

\item{seed}{an (optional) seeding number.}

\item{parRidge}{an (optional) list of parameters passed on to glmnet for the ridge
stage. Format argument = value.}

\item{parLasso}{an (optional) list of parameters passed on to glmnet for the LASSO
stage. Format argument = value.}
}
\value{
a banditGlmnet object.
}
\description{
Estimate a banditGlmnet model
}
