% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banditMethods.R
\name{coef.bandit}
\alias{coef.bandit}
\title{Export model coefficients.}
\usage{
coef.bandit(object, modelId = NULL, as.matrix = FALSE)
}
\arguments{
\item{object}{a bandit object}

\item{modelId}{(optional) a vector of ids of training jobs.}

\item{restricted}{a character in 'ridge', 'lasso' for the stage of the desired model.}
}
\value{
Coefficients extracted from the model modelId. The default returns the ridge
coefficients of the last model.
}
\description{
Export model coefficients.
}

